requirements.txt
openai==1.6.1
python-dotenv==1.0.0
uvicorn==0.25.0
fastapi==0.105.0
sentence-transformers==2.2.2
llama-index==0.9.21
pymongo==4.6.1
python-multipart==0.0.6
gunicorn==21.2.0
pypdf==3.17.4
boto3==1.34.23
textblob==0.9.0
google-cloud-aiplatform